Improve variable names and logical flow in Command.java
Variable names are ambiguous, and code is unreadable. Due
to having too many types of inputs that users
can supply, variables are carelessly named.

As more commands are being added, and new requirements
are being imposed on the commands, the code must be
easily maintainable by making it easy to understand
the current behaviour. This way, when adding new
requirements, bugs will not be accidentally introduced.

Let's
* rename the variables to be less ambiguous
* different case blocks to not share any variable names
if they are not declared before the switch case block.

If the variable names are being declared in an earlier
block, and being used in a later block, this means
that the scope of the variable can be further narrowed
down to be more specific. This increases readability
and reduces ambiguity.